+++ src/collective/collectionfilter/interfaces.py
@@ -178,6 +178,32 @@ class ICollectionSearchSchema(ICollectionFilterBaseSchema):
     """
 
 
+class ICollectionFilterResultListSort(ICollectionFilterBaseSchema):
+    """Schema for the result list sorting.
+    """
+
+    sort_on = schema.Choice(
+        title=_('label_sort_on', u'Enabled sort indexes'),
+        description=_(
+            'help_sort_on',
+            u'Select the indexes which can be sorted on.'
+        ),
+        required=True,
+        vocabulary='collective.collectionfilter.SortOnIndexes',
+    )
+
+    input_type = schema.Choice(
+        title=_('label_input_type', u'Input Type'),
+        description=_(
+            'help_input_type',
+            u'Select how the UI of the collection filter should be rendered. '
+            u'Wether as links, as checkboxes and radiobuttons or checkboxes and dropdowns.'  # noqa
+        ),
+        required=True,
+        vocabulary='collective.collectionfilter.InputType',
+    )
+
+
 class IGroupByCriteria(Interface):
     """Interface for the GroupByCriteria utility.